Moved customers app into users app
