+++ cloudkitchen/settings/base.py
@@ -46,7 +46,6 @@ USER_APPS = [
     'orders.apps.OrdersConfig',
     'kitchen.apps.KitchenConfig',
     'diners.apps.DinersConfig',
-    'customers.apps.CustomersConfig',
     'api.apps.ApiConfig',
 ]
