+++ cloudkitchen/templates/dabbawala/nav.html
@@ -16,7 +16,7 @@
       <li><a href="{% url 'products:menu'%}">Menú</a></li>
       <li><a href="{% url 'branch_offices:contact' %}">Contacto</a></li>
       <li><a href="" class="button nob" data-type="zoomin"><span>Inicia Sesión</span></a></li>
-      <li><a href="{% url 'customers:register' %}" class="button fit">Resgistrate</a></li>
+      <li><a href="{% url 'users:register' %}" class="button fit">Resgistrate</a></li>
     </ul>
   </nav>
 
@@ -30,7 +30,7 @@
     </div>
     <a href="" title=""></a>
     <div class="menres">
-      <a href="{% url 'customers:register' %}" class="button"><span>Resgistrate </span></a>
+      <a href="{% url 'users:register' %}" class="button"><span>Resgistrate </span></a>
     </div>
   </nav>
 </header>
@@ -48,7 +48,7 @@
     <div class="colm2">
       <h1>Registro</h1>
       <p>Si aún no te has registrado, te invitamos a que visites nuestra sección de registro y disfrutes de todos los beneficios que tu cuenta te otorga.</p>
-      <a href="{% url 'customers:register' %}" class="button"><span>Resgistrate </span></a>
+      <a href="{% url 'users:register' %}" class="button"><span>Resgistrate </span></a>
     </div>
     <span class="closeres">X</span>
   </div>
