+++ users/templates/register/customers_list.html
