+++ users/templates/register/new_customer.html
