+++ users/templates/register/register.html
@@ -12,7 +12,7 @@
 {% endblock stylesheets%}
 
 {% block content %}
-<div id="main1"> 
+<div id="main1">
   <img class="bord2" src="{% static 'images/image/reverse.png' %}" alt="Borde">
   <h2>bienvenido a </h2>
   <h1>Dabbawala</h1>
@@ -93,7 +93,7 @@
               <span id="msj-alert" class="alertmap">Arrastra el marcador a tu posición exacta.</span>
             </div>
           </div>
-          <div id="map" class=""></div> 
+          <div id="map" class=""></div>
         </div>
       </div>
     </form>
@@ -134,7 +134,7 @@
 
   function setLatLng(lat, lng){
     $('#id_latitude').val(lat);
-    $('#id_longitude').val(lng); 
+    $('#id_longitude').val(lng);
   }
 
   function openInfoWindow(marker) {
@@ -155,11 +155,11 @@
   }
 
   function addMarkerLsteners (marker) {
-    google.maps.event.addListener(marker, 'dragend', function(){ 
-      openInfoWindow(marker); 
+    google.maps.event.addListener(marker, 'dragend', function(){
+      openInfoWindow(marker);
     });
-    google.maps.event.addListener(marker, 'click', function(){ 
-      openInfoWindow(marker); 
+    google.maps.event.addListener(marker, 'click', function(){
+      openInfoWindow(marker);
     });
   }
 
@@ -210,7 +210,7 @@
 
     function search(e) {
       var code = (e.keyCode ? e.keyCode : e.which);
-      if(code == 13) { 
+      if(code == 13) {
         var address = $('#id_address').val();
         var geocoder = new google.maps.Geocoder();
         geocoder.geocode({'address': address}, geocodeResult);
@@ -232,9 +232,9 @@
         drawMap(optionsMap);
         map.fitBounds(results[0].geometry.viewport);
 
-        var markerOptions = { 
+        var markerOptions = {
           position: results[0].geometry.location,
-          draggable: true, 
+          draggable: true,
           animation: google.maps.Animation.DROP,
         }
 
@@ -279,14 +279,14 @@
       let userJson = JSON.stringify(new_user);
 
       $.ajax({
-        url: "{% url 'customers:register' %}",
+        url: "{% url 'users:register' %}",
         type: 'POST',
         data: {
           csrfmiddlewaretoken: '{{ csrf_token }}',
           'newuser': userJson,
           'type': 'regist_user'
         },
-        traditional: true, 
+        traditional: true,
         datatype: 'jsonp',
       });
