+++ users/templates/register/thanks.html
@@ -3,7 +3,7 @@
 {% block alert %}
 <script type="text/javascript" charset="utf-8">
 	function rel(){
-		        window.location.href = "{% url 'customers:register' %}";
+		        window.location.href = "{% url 'users:register' %}";
 
 	}
 	function showAlert() {
@@ -13,8 +13,8 @@
 			type: "success",
 			timer: 2500,
 			showConfirmButton: false
-		});  
-		setTimeout(rel, 2100);	
+		});
+		setTimeout(rel, 2100);
 	}
 	$(function() {
 		showAlert();
