+++ users/views.py
@@ -1,5 +1,7 @@
+import json
+from django.conf import settings
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render, redirect
 
 from django.contrib.auth import authenticate
@@ -7,7 +9,9 @@ from django.contrib.auth import login as login_django
 from django.contrib.auth import logout as logout_django
 
 from users.forms import UserForm
-from users.models import UserMovements
+from users.models import UserMovements, CustomerProfile
+
+from .forms import CustomerProfileForm
 
 from cloudkitchen.settings.base import PAGE_TITLE
 
@@ -61,7 +65,7 @@ def login(request):
             user = authenticate(username=username_login, password=password_login)
 
             if user is not None:
-                login_django(request, user)                   
+                login_django(request, user)
                 login_check(user.username)
                 return redirect('sales:sales')
 
@@ -95,15 +99,108 @@ def login_register(request):
     'titie' : title,
     'objects' : objects
     }
-    return render(request, template, context)    
+    return render(request, template, context)
 
 
-def login_check(user):       
-    movement = UserMovements.objects.create(category='LogIn',user=user)    
+def login_check(user):
+    movement = UserMovements.objects.create(category='LogIn',user=user)
     movement.save()
 
 
 @login_required(login_url='users:login')
-def logout_check(user):       
-    movement = UserMovements.objects.create(category='LogOut',user=user)    
+def logout_check(user):
+    movement = UserMovements.objects.create(category='LogOut',user=user)
     movement.save()
+
+
+# -------------------------------------  Customers -------------------------------------
+def new_customer(request):
+    form_customer = CustomerProfileForm(request.POST or None)
+    if request.method == 'POST':
+        print(request.POST)
+        if form_customer.is_valid():
+            customer = form_customer.save(commit=False)
+            customer.save()
+            return redirect('customers:thanks')
+
+    template = 'register/new_customer.html'
+    title = 'Dabbawala - Bienvenido a Dabbawala. Registrare y obtén un desayuno gratis. '
+    context = {
+        'form_customer': form_customer,
+        'title': title,
+    }
+
+    return render(request, template, context)
+
+
+def register(request):
+    form_customer = CustomerProfileForm(request.POST or None)
+    if request.method == 'POST':
+        if form_customer.is_valid():
+            customer = form_customer.save(commit=False)
+            customer.save()
+            return redirect('customers:thanks')
+
+    template = 'register/register.html'
+    title = 'Bienvenido a Dabbawala.'
+    context = {
+        'form_customer': form_customer,
+        'title': title,
+    }
+    if request.method == 'POST':
+        if 'newuser' in request.POST:
+            if request.POST['type'] == 'regist_user':
+                new_user = request.POST['newuser']
+                if settings.DEBUG:
+                    print(new_user)
+
+    return render(request, template, context)
+
+
+
+def thanks(request):
+    if request.method == 'POST':
+        form = CustomerProfileForm(request.POST, request.FILES)
+        if form.is_valid():
+            customer = form.save(commit=False)
+            customer.save()
+            return redirect('customers:new_customer')
+    else:
+        form = CustomerProfileForm()
+
+    template = 'register/thanks.html'
+    title = 'Dabbawala - Registro de clientes'
+
+    context = {
+        'form': form,
+        'title': title,
+    }
+
+    return render(request, template, context)
+
+
+
+@login_required(login_url='users:login')
+def customers_list(request):
+    if request.method == 'POST':
+        customer_json_object = json.loads(request.POST.get('customer'))
+
+        customer_object = CustomerProfile.objects.get(id=customer_json_object['id'])
+        customer_object.first_dabba = True
+        customer_object.save()
+        data = {
+            'status': 'ready'
+        }
+        return JsonResponse(data)
+
+    template = 'register/customers_list.html'
+    customers = CustomerProfile.objects.all().order_by('first_dabba')
+    title = 'Clientes registrados'
+
+    context = {
+        'title': title,
+        'page_title': PAGE_TITLE,
+        'customers': customers,
+    }
+
+    return render(request, template, context)
